+++ src/screens/Home/styles.tsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+export const HomeContainer = styled.div`
+  align-items: center;
+  display: flex;
+  flex-direction: column;
+  height: 100vh;
+  justify-content: center;
+  width: 100vw;
+`;