+++ src/screens/UserInfo/styles.tsx
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+export const UserInfoContainer = styled.div`
+  height: 100vh;
+  width: 100vw;
+  padding: 2.2rem 1.6rem;
+  max-width: 1100px;
+  margin: 0 auto;
+`;
+
+export const UserInfoContent = styled.div`
+  display: flex;
+  margin-top: 3.1rem;
+`;