+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Added 
+- [Adding Global Style, fonts, favicon and page title - issue #5](https://github.com/rssviana/foribus-oraculi/issues/5)
+- [Adding Header Component - issue #5](https://github.com/rssviana/foribus-oraculi/issues/5)
+- [Adding Hero Component - issue #5](https://github.com/rssviana/foribus-oraculi/issues/5)
+- [Adding Repositories Component - issue #5](https://github.com/rssviana/foribus-oraculi/issues/5)
+- [Adding Sidebar Component - issue #5](https://github.com/rssviana/foribus-oraculi/issues/5)
+- [Adding Home Page - issue #5](https://github.com/rssviana/foribus-oraculi/issues/5)
+- [Adding UserInfo Page - issue #5](https://github.com/rssviana/foribus-oraculi/issues/5)
+- [Adding NotFound Page - issue #5](https://github.com/rssviana/foribus-oraculi/issues/5)
+- [Adding Routes - issue #5](https://github.com/rssviana/foribus-oraculi/issues/5)
+
+
+
 ## [0.0.1] - 10-12-2020
 
 ### Added